+++ clojure-mode.el
@@ -196,9 +196,9 @@ Out-of-the box `clojure-mode' understands lein, boot, gradle,
           (and (listp value)
                (cl-every 'stringp value))))
 
-(defcustom clojure-elided-prefixes-ns
-  '()
-  "A list of additional regexp prefixes that are elided when formulating ns."
+(defcustom clojure-directory-prefixes
+  '("\\`clj[scx]?\\.")
+  "A list of directory prefixes used by clojure-expected-ns to formulate correct ns."
   :type '(repeat string)
   :package-version '(clojure-mode . "5.0.0")
   :safe (lambda (value)
@@ -1770,9 +1770,8 @@ If PATH is nil, use the path to the file backing the current buffer."
          (sans-file-sep (mapconcat 'identity (cdr (split-string sans-file-type "/")) "."))
          (sans-underscores (replace-regexp-in-string "_" "-" sans-file-sep)))
     ;; Drop prefix from ns for projects with structure src/{clj,cljs,cljc}
-    (cl-reduce (lambda (a x)
-                 (replace-regexp-in-string x "" a))
-               (add-to-list 'clojure-elided-prefixes-ns "\\`clj[scx]?\\.")
+    (cl-reduce (lambda (a x) (replace-regexp-in-string x "" a))
+               'clojure-directory-prefixes
                :initial-value sans-underscores)))
 
 (defun clojure-insert-ns-form-at-point ()